==> Source Code/01_data_processing.ipynb <==
# Data Processing

## Created By: Aldo Iturrios and Alyssa Wisk

In this notebook, we'll be reading in all txt files, and stacking each file into one single (rectangular) dataframe). 

import numpy as np
import pandas as pd
import os
from sklearn.linear_model import LinearRegression

Below, we'll be stacking all txt files into one single dataframe. There are two other important things happening in the for loop below:
* If a patient has more than 1 measurement recorded for a single measure in a one half hour interval, we take the average of all those measurements in that half hour interval and list the average. Ex. If in hour 1:30, patient had blood pressure taken 3 times, we take the average of those 3 measurements, and that is the number that ends up in the data frame for that patient.
* We will also be taking a linear regression for each measure (across all time) for each patient. Coefficent associated with this measure will serve as a summary of the trend for that measure over time.

# Set folder where all data is stored
folder = "data/x_all/"

# Construct empty dataframe to store all data
train_data = pd.DataFrame()

for i, patient in enumerate(os.listdir(folder)):
    
    # Foundation
    patient_id = patient[:-4]
    file = os.path.join(folder, patient)
    df = pd.read_csv(file, delimiter = ",", dtype={'Time': str, 'Variable':str})
    
    # Regression for each variable
    linear_regressor = LinearRegression() 
    df_reg = df.copy().dropna()
    df_reg['Time2'] = df_reg['Time'].apply(lambda x: int(x[0:2])*60 + int(x[3:5]))
    
    vars = df_reg.Variable.unique()
    dict_reg = {}

    for v in vars:
        data = df_reg[df_reg['Variable'] == v]
        try:
            reg = linear_regressor.fit(data['Time2'].to_numpy().reshape(-1, 1), data['Value'].to_numpy())
            dict_reg[v] = reg.coef_
        except:
            pass   
            dict_reg[v] = 0
            
    

    # Form for final data
    df1 = df.melt(id_vars=['Variable', 'Time'], value_vars=['Value'])
    df1['Feature'] = df1['Time'] + "_" + df1['Variable']
    df2 = df1.groupby('Feature')['value'].mean().to_frame().T
    df2 = df2.set_index(pd.Series([patient_id]))
    
    patient_data = pd.concat([df2, pd.DataFrame(dict_reg).set_index(pd.Series([patient_id]))], axis=1)
    
    if i == 0:
        train_data = patient_data
    else:
        train_data = train_data.append(patient_data, sort=False)

train_data.head(5)

## Further adjustments
* Remove Time stamp from 4 initial variables
* Name index into "id" (to match train_outcome.csv file)
* Order columns

train_data_dc = train_data.copy()
train_data_dc = train_data_dc.drop(columns=["AdmissionType", "Age", "Gender", "RecordID"])
train_data_dc = train_data_dc.sort_values(by = ["00:00_RecordID"]).reset_index()
train_data_dc = train_data_dc.rename(columns={"index": "id", 
                              "00:00_AdmissionType": "AdmissionType",
                              "00:00_Age": "Age",
                              "00:00_Gender": "Gender",
                              "00:00_RecordID": "RecordID"})

train_data_dc = train_data_dc.reindex(sorted(train_data_dc.columns), axis=1)
cols_at_beg = ["id", "AdmissionType", "Age", "Gender", "RecordID"]
train_data_dc = train_data_dc[[c for c in cols_at_beg if c in train_data_dc] 
                              + [c for c in train_data_dc if c not in cols_at_beg]]
train_data_dc.sort_values(by = ["RecordID"])
train_data_dc.head(5)

## Save Data Frame

train_data_dc.to_csv("data/patient_data/patient_dataframe.csv", index=False)
